[#1500] Mock pylons context object on paster commands
When calling a paster command, we create and register a Pylons context
object (c), and add to site user to it (c.user and c.auth_user_object).
This ensures that the auth will work in exactly the same way as in a web
request. Otherwise you can find problems like the harvest and DataPusher
incompatibility.

We already did a similar thing registering a MockTranslator for the
thread.